pyproject.toml
[build-system]
requires = ["setuptools>=42", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "gmail-mcp"
version = "0.1.0"
description = "Gmail MCP for Claude"
requires-python = ">=3.10"
dependencies = [
    "google-api-python-client>=2.0.0",
    "google-auth-httplib2>=0.1.0",
    "google-auth-oauthlib>=0.4.0",
    "httpx>=0.28.1",
    "mcp>=0.1.0",
    "python-dotenv>=1.0.0",
]

[project.scripts]
gmail-mcp = "gmail_mcp.main:main"
